[Test] 상품 테스트 API 별로 분리
